+++ packages/slate-plugins/src/deserializers/deserialize-html/__tests__/deserializeHTMLToDocument/empty-span.spec.tsx
@@ -16,5 +16,10 @@ const output = (
 ) as any;
 
 it('should be', () => {
-  expect(deserializeHTMLToDocument(plugins)(body)).toEqual(output);
+  expect(
+    deserializeHTMLToDocument({
+      plugins,
+      element: body,
+    })
+  ).toEqual(output);
 });