+++ packages/slate-plugins/src/deserializers/deserialize-html/__tests__/deserializeHTMLToDocument/two-div.spec.tsx
@@ -19,5 +19,10 @@ const output = (
 ) as any;
 
 it('should have the break line', () => {
-  expect(deserializeHTMLToDocument(input1)(input2)).toEqual(output);
+  expect(
+    deserializeHTMLToDocument({
+      plugins: input1,
+      element: input2,
+    })
+  ).toEqual(output);
 });