+++ packages/slate-plugins/src/deserializers/deserialize-html/__tests__/deserializeHTMLToMarks/children-text.spec.tsx
@@ -9,7 +9,7 @@ import {
 
 const input: DeserializeMarksProps = {
   plugins: [BoldPlugin()],
-  el: document.createElement('strong'),
+  element: document.createElement('strong'),
   children: <fragment>test</fragment>,
 } as any;